Refactor landing page, extrude functions and components
